requirements.txt
django-import-export
psycopg2-binary
Django<=3.2 # https://docs.django-cms.org/en/latest/#django-python-compatibility-table

# Django CMS
django-cms
djangocms-file
djangocms-link
djangocms-text-ckeditor
djangocms-picture
djangocms-redirect
